GH-87 Refactoring of tests
- Extract temporary file handling
- Extract subscription management
