+++ rosbag2_tests/CMakeLists.txt
@@ -32,6 +32,7 @@ if(BUILD_TESTING)
 
   ament_add_gmock(test_rosbag2_record_end_to_end
     test/test_rosbag2_record_end_to_end.cpp
+    test/test_memory_management.cpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
     ENV ROS_DOMAIN_ID=42) # make sure this test does not interfer with transport tests
   if(TARGET test_rosbag2_record_end_to_end)
@@ -44,6 +45,7 @@ if(BUILD_TESTING)
 
   ament_add_gmock(test_rosbag2_play_end_to_end
     test/test_rosbag2_play_end_to_end.cpp
+    test/test_memory_management.cpp
     WORKING_DIRECTORY ${CMAKE_CURRENT_BINARY_DIR}
     ENV ROS_DOMAIN_ID=42) # make sure this test does not interfer with transport tests
   if(TARGET test_rosbag2_play_end_to_end)
